feat(processing): add revision manager class
